Make sure we're using branch `main`
PR #51 updated gibo to use the new `main` branch in the github/gitignore repo, rather than the old `master` branch. But if a user had installed gibo already, the local clone of github/gitignore stays on the `master` branch. This commit fixes that, by making sure every time we run `gibo update` that we're using the `main` branch.