Simplify and unify go versioning
Since gibo is a distributed binary and not used as a dependency by
others, there's no need to use a lower `go` version in go.mod. I updated
the `go` version to match the `toolchain` version, then when I ran `go
mod tidy` it removed the now-redundant `toolchain` line.

Also udpate the Dockerfile to use the same version.